Change cpu_idle_func() to double
git-svn-id: http://svn.research.cc.gatech.edu/kaos/evpath/trunk@7503 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ metrics.c
@@ -145,7 +145,7 @@ unsigned long total_jiffies_func ( void )
 }   
 
  
-float cpu_idle_func ( void )
+double cpu_idle_func ( void )
 {
     char *p;
     static float val; 
@@ -178,7 +178,7 @@ stone_type stone;
 cod_parse_context context;
 {
     static char extern_string[] = "\
-		float cpu_idle_func();\n\
+		double cpu_idle_func();\n\
 		int num_cpustates_func();\n";
 
     static cod_extern_entry externs[] = {
